# api/gem/internal/gen/server.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package gen

import (
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	strictgin "github.com/oapi-codegen/runtime/strictmiddleware/gin"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// ジェムの残高取得API
	// (GET /v1/currencies/balances)
	GetGemBalance(c *gin.Context)
	// ジェムの購入API
	// (PUT /v1/currencies/purchases)
	UpdateGemPurchase(c *gin.Context)
	// ヘルスチェックAPI
	// (GET /v1/healthcheck)
	Healthcheck(c *gin.Context)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// GetGemBalance operation middleware
func (siw *ServerInterfaceWrapper) GetGemBalance(c *gin.Context) {

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetGemBalance(c)
}

// UpdateGemPurchase operation middleware
func (siw *ServerInterfaceWrapper) UpdateGemPurchase(c *gin.Context) {

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.UpdateGemPurchase(c)
}

// Healthcheck operation middleware
func (siw *ServerInterfaceWrapper) Healthcheck(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.Healthcheck(c)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/v1/currencies/balances", wrapper.GetGemBalance)
	router.PUT(options.BaseURL+"/v1/currencies/purchases", wrapper.UpdateGemPurchase)
	router.GET(options.BaseURL+"/v1/healthcheck", wrapper.Healthcheck)
}

type BadRequestResponse struct {
}

type ForbiddenResponse struct {
}

type InternalServerErrorResponse struct {
}

type NotFoundResponse struct {
}

type UnauthorizedResponse struct {
}

type GetGemBalanceRequestObject struct {
}

type GetGemBalanceResponseObject interface {
	VisitGetGemBalanceResponse(w http.ResponseWriter) error
}

type GetGemBalance200JSONResponse Balance

func (response GetGemBalance200JSONResponse) VisitGetGemBalanceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetGemBalance401Response = UnauthorizedResponse

func (response GetGemBalance401Response) VisitGetGemBalanceResponse(w http.ResponseWriter) error {
	w.WriteHeader(401)
	return nil
}

type GetGemBalance404Response = NotFoundResponse

func (response GetGemBalance404Response) VisitGetGemBalanceResponse(w http.ResponseWriter) error {
	w.WriteHeader(404)
	return nil
}

type GetGemBalance500Response = InternalServerErrorResponse

func (response GetGemBalance500Response) VisitGetGemBalanceResponse(w http.ResponseWriter) error {
	w.WriteHeader(500)
	return nil
}

type UpdateGemPurchaseRequestObject struct {
	Body *UpdateGemPurchaseJSONRequestBody
}

type UpdateGemPurchaseResponseObject interface {
	VisitUpdateGemPurchaseResponse(w http.ResponseWriter) error
}

type UpdateGemPurchase201JSONResponse PurchaseResponse

func (response UpdateGemPurchase201JSONResponse) VisitUpdateGemPurchaseResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type UpdateGemPurchase400Response = BadRequestResponse

func (response UpdateGemPurchase400Response) VisitUpdateGemPurchaseResponse(w http.ResponseWriter) error {
	w.WriteHeader(400)
	return nil
}

type UpdateGemPurchase401Response = UnauthorizedResponse

func (response UpdateGemPurchase401Response) VisitUpdateGemPurchaseResponse(w http.ResponseWriter) error {
	w.WriteHeader(401)
	return nil
}

type UpdateGemPurchase403Response = ForbiddenResponse

func (response UpdateGemPurchase403Response) VisitUpdateGemPurchaseResponse(w http.ResponseWriter) error {
	w.WriteHeader(403)
	return nil
}

type UpdateGemPurchase404Response = NotFoundResponse

func (response UpdateGemPurchase404Response) VisitUpdateGemPurchaseResponse(w http.ResponseWriter) error {
	w.WriteHeader(404)
	return nil
}

type UpdateGemPurchase500Response = InternalServerErrorResponse

func (response UpdateGemPurchase500Response) VisitUpdateGemPurchaseResponse(w http.ResponseWriter) error {
	w.WriteHeader(500)
	return nil
}

type HealthcheckRequestObject struct {
}

type HealthcheckResponseObject interface {
	VisitHealthcheckResponse(w http.ResponseWriter) error
}

type Healthcheck200JSONResponse HealthCheck

func (response Healthcheck200JSONResponse) VisitHealthcheckResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// ジェムの残高取得API
	// (GET /v1/currencies/balances)
	GetGemBalance(ctx *gin.Context, request GetGemBalanceRequestObject) (GetGemBalanceResponseObject, error)
	// ジェムの購入API
	// (PUT /v1/currencies/purchases)
	UpdateGemPurchase(ctx *gin.Context, request UpdateGemPurchaseRequestObject) (UpdateGemPurchaseResponseObject, error)
	// ヘルスチェックAPI
	// (GET /v1/healthcheck)
	Healthcheck(ctx *gin.Context, request HealthcheckRequestObject) (HealthcheckResponseObject, error)
}

type StrictHandlerFunc = strictgin.StrictGinHandlerFunc
type StrictMiddlewareFunc = strictgin.StrictGinMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// GetGemBalance operation middleware
func (sh *strictHandler) GetGemBalance(ctx *gin.Context) {
	var request GetGemBalanceRequestObject

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetGemBalance(ctx, request.(GetGemBalanceRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetGemBalance")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(GetGemBalanceResponseObject); ok {
		if err := validResponse.VisitGetGemBalanceResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// UpdateGemPurchase operation middleware
func (sh *strictHandler) UpdateGemPurchase(ctx *gin.Context) {
	var request UpdateGemPurchaseRequestObject

	var body UpdateGemPurchaseJSONRequestBody
	if err := ctx.ShouldBind(&body); err != nil {
		ctx.Status(http.StatusBadRequest)
		ctx.Error(err)
		return
	}
	request.Body = &body

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.UpdateGemPurchase(ctx, request.(UpdateGemPurchaseRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "UpdateGemPurchase")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(UpdateGemPurchaseResponseObject); ok {
		if err := validResponse.VisitUpdateGemPurchaseResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// Healthcheck operation middleware
func (sh *strictHandler) Healthcheck(ctx *gin.Context) {
	var request HealthcheckRequestObject

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.Healthcheck(ctx, request.(HealthcheckRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "Healthcheck")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(HealthcheckResponseObject); ok {
		if err := validResponse.VisitHealthcheckResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}