# api/gem/spec/openapi.yaml
openapi: 3.0.0
info:
  title: async serverless application gem API
  description: async serverless application gem API
  version: 1.0.0
servers:
  - url: http://localhost:{port}/gem/
    description: dev
    variables:
      port:
        default: '8081'
  - url: https://api.<domain-name>/gem/
    description: stg
security:
  - bearerAuth: []
paths:
  /v1/healthcheck:
    get:
      tags:
        - healthcheck
      summary: ヘルスチェックAPI
      description: APIが正常に動作しているかを確認します。
      operationId: healthcheck
      responses:
        '200':
          description: API is up and running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheck'
      security: []
  /v1/currencies/balances:
    get:
      tags:
        - v1
        - currencies
      summary: ジェムの残高取得API
      description: 現在保有しているジェムの残高を取得する
      operationId: getGemBalance
      responses:
        '200':
          description: 現在保有しているジェムの残高取得に成功しました。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Balance'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/currencies/purchases:
    put:
      tags:
        - v1
        - currencies
      summary: ジェムの購入API
      description: ジェムを購入し、残高に加算する
      operationId: updateGemPurchase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseRequest'
      responses:
        '201':
          description: ジェムの購入に成功しました。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthCheck:
      title: HealthCheck
      type: object
      description: HealthCheckのレスポンス
      required:
        - message
      properties:
        message:
          description: メッセージ。通常はok
          type: string
      example:
        message: ok
    Balance:
      type: object
      required:
        - all
        - paid
        - free
      properties:
        all:
          description: 保有している全ての通貨の残高
          type: integer
          format: uint32
          example: 1500
        paid:
          description: 保有している有償通貨の残高
          type: integer
          format: uint32
          example: 1000
        free:
          description: 保有している無償通貨の残高
          type: integer
          format: uint32
          example: 500
    PurchaseRequest:
      type: object
      required:
        - gem_id
        - quantity
      properties:
        gem_id:
          description: 購入するジェムのID
          type: integer
          format: uint32
          example: 10001010
        quantity:
          description: 購入するジェムの数量
          type: integer
          format: uint32
          example: 1000
    PurchaseResponse:
      type: object
      required:
        - balance
        - transaction_id
      properties:
        balance:
          description: 購入後のジェム残高
          type: integer
          format: uint32
          example: 100000
        transaction_id:
          description: トランザクションID
          type: string
          example: tx_purchase_001
  responses:
    Unauthorized:
      description: Unauthorized
    NotFound:
      description: Not Found
    InternalServerError:
      description: Internal Server Error
    BadRequest:
      description: Bad Request
    Forbidden:
      description: Forbidden